Rename valve file to match
